split out mute button into its own component
