+++ src/scenes/ExplainerScene/ExplainerScene.jsx
@@ -1,16 +1,8 @@
 // This page is required to make audio autoplay because of google policies.
-
-import { useContext } from "react";
+import MuteButton from "../../components/MuteButton/MuteButton";
 import "./ExplainerScene.css";
-import { EnableSoundContext } from "../../App";
 
 const ExplainerScene = ({ nextScene }) => {
-  const {isMuted, setIsMuted} = useContext(EnableSoundContext);
-
-  const toggleMute = () => {
-    setIsMuted(!isMuted);
-  };
-
   return (
     <div style={{ background: "black", width: "100%", height: "100%", display: "flex", alignItems: "center", justifyContent: "center" }}>
       <center style={{ maxWidth: "650px" }}>
@@ -25,17 +17,11 @@ const ExplainerScene = ({ nextScene }) => {
           I needed to create this page in order to have audio autoplay because of google's autoplay policy so keep your volume on!
         </i>
         <br />
-        <button className="button-18" onClick={() => nextScene()}>Enter</button>
+        <div style={{ display: "flex", alignItems: "center", justifyContent: "center", marginTop: "1rem" }}>
+          <MuteButton />
+          <button className="button-18" style={{marginTop: "0px", marginLeft: "1.25rem"}} onClick={() => nextScene()}>Enter</button>
+        </div>
       </center>
-      <button style={{ backgroundColor: 'transparent', border: '1px solid white', borderRadius: '50%', width: '2rem', height: '2rem', marginLeft: '0.5rem' }} onClick={toggleMute}>
-          <svg style={{ fill: 'white', width: '1.5rem', height: '1.5rem', margin: 'auto' }} viewBox="0 0 24 24">
-            {isMuted ? (
-              <path d="M7 9v6h4l5 5V4l-5 5H7z" />
-            ) : (
-              <path d="M4 9v6h4l5 5V4l-5 5H4z" />
-            )}
-          </svg>
-        </button>
     </div>
   )
 }
